+++ scheduler/src/main/java/org/apache/mesos/elasticsearch/scheduler/state/ClusterState.java
@@ -6,6 +6,7 @@ import org.apache.mesos.Protos.TaskInfo;
 import org.apache.mesos.elasticsearch.scheduler.Task;
 import org.apache.mesos.elasticsearch.scheduler.TaskInfoFactory;
 
+import javax.validation.constraints.NotNull;
 import java.io.IOException;
 import java.io.NotSerializableException;
 import java.security.InvalidParameterException;
@@ -23,7 +24,10 @@ public class ClusterState {
     private FrameworkState frameworkState;
     private TaskInfoFactory taskInfoFactory;
 
-    public ClusterState(SerializableState zooKeeperStateDriver, FrameworkState frameworkState, TaskInfoFactory taskInfoFactory) {
+    public ClusterState(@NotNull SerializableState zooKeeperStateDriver, @NotNull FrameworkState frameworkState, @NotNull TaskInfoFactory taskInfoFactory) {
+        if (zooKeeperStateDriver == null || frameworkState == null) {
+            throw new NullPointerException();
+        }
         this.zooKeeperStateDriver = zooKeeperStateDriver;
         this.frameworkState = frameworkState;
         this.taskInfoFactory = taskInfoFactory;
@@ -194,7 +198,7 @@ public class ClusterState {
         }
     }
 
-    private String getKey() throws NotSerializableException {
+    private String getKey() {
         return frameworkState.getFrameworkID().getValue() + "/" + STATE_LIST;
     }
 }